requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
lxml==4.9.3 ; sys_platform != 'linux' or python_version < '3.12'
fake-useragent==1.4.0
selenium==4.15.0
# For Linux/Python 3.12+ use system packages
